+++ lib/config/config.go
@@ -5,7 +5,7 @@ import (
   "sort"
   "strings"
 
-  "github.com/Originate/gt/cmd/util"
+  "github.com/Originate/gt/lib/util"
 )
