+++ lib/git/status.go
@@ -4,7 +4,7 @@ import (
   "fmt"
   "os"
 
-  "github.com/Originate/gt/cmd/util"
+  "github.com/Originate/gt/lib/util"
 )
