+++ lib/steps/create_and_checkout_branch_step.go
@@ -1,8 +1,8 @@
 package steps
 
 import (
-  "github.com/Originate/gt/cmd/config"
-  "github.com/Originate/gt/cmd/script"
+  "github.com/Originate/gt/lib/config"
+  "github.com/Originate/gt/lib/script"
 )
