+++ lib/steps/fetch_upstream_step.go
@@ -1,7 +1,7 @@
 package steps
 
 import (
-  "github.com/Originate/gt/cmd/script"
+  "github.com/Originate/gt/lib/script"
 )
