+++ lib/steps/rebase_tracking_branch_step.go
@@ -1,7 +1,7 @@
 package steps
 
 import (
-  "github.com/Originate/gt/cmd/git"
+  "github.com/Originate/gt/lib/git"
 )
