+++ lib/steps/run.go
@@ -4,7 +4,7 @@ import (
   "fmt"
   "os"
 
-  "github.com/Originate/gt/cmd/git"
+  "github.com/Originate/gt/lib/git"
 
   "github.com/fatih/color"
 )
